requirements.txt
# requirements.txt
albucore
albumentations
annotated-types
anyio
astor
attrs
beautifulsoup4
bidict
Cerberus
certifi
cffi
charset-normalizer
colorama
colorthief
contourpy
cryptography
cv2ext
cycler
Cython
darkdetect
decorator
dill
filelock
fire
fonttools
fsspec
h11
httpcore
httpx
idna
imageio
imgaug
Jinja2
joblib
jsonschema
jsonschema-specifications
kiwisolver
lazy_loader
lmdb
loguru
lxml
MarkupSafe
marshmallow
matplotlib
MouseInfo
mpmath
msgpack
mss
networkx
numpy
opencv-contrib-python
opencv-python
opencv-python-headless
opt-einsum
packaging
paddleocr
paddlepaddle
paddlepaddle-gpu
pandas
pillow
protobuf
psutil
py-cpuinfo
PyAutoGUI-ng
pyclipper
pycparser
pydantic
pydantic_core
pydirectinput_rgx
PyGetWindow
PyMonCtl
PyMsgBox
pyparsing
pyperclip
PyQt-builder
PyRect
PyScreeze
PySide6
PySide6_Addons
PySide6_Essentials
PySide6-Fluent-Widgets
PySideSix-Frameless-Window
python-dateutil
python-docx
pytweening
pytz
pywin32
pywin32-ctypes
PyWinBox
PyWinCtl
PyYAML
RapidFuzz
referencing
requests
rpds-py
ruamel.yaml
ruamel.yaml.clib
scikit-image
scikit-learn
scipy
seaborn
shapely
shiboken6
sip
six
sniffio
soupsieve
sympy
termcolor
threadpoolctl
tifffile
toml
tomli
tqdm
typing_extensions
tzdata
ultralytics
ultralytics-thop
urllib3
watchdog
win32-setctime
windows-capture
